pyproject.toml
[project]
name = "originbluy-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-community>=0.3.27",
    "langchain[openai]>=0.3.27",
    "pypdf>=5.9.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
